+++ ui/src/app/core/accessories/accessories.service.ts
@@ -35,6 +35,7 @@ export class AccessoriesService {
 
   private hiddenTypes = [
     'InputSource',
+    'CameraRTPStreamManagement',
   ];
 
   constructor(